Implements `list` and `reduce`
I had to add isEmpty() method to Seq interface which I don't quite
like but I didn't see another way to identify empty Seqs. Calling
rest on an empty Seq returns an empty Seq. Calling rest on the a Seq
of 1 element in it returns an empty seq also.

One can't check if we are at the final Cell, since there's still a
value in it. That could very well be null, as we can have a Seq
containing null values.